--- app/controllers/api/v1/base_controller.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-# rubocop:disable Rails/ApplicationController
-
-module Api
-  module V1
-    class BaseController < ActionController::Base
-    end
-  end
-end
-
-# rubocop:enable Rails/ApplicationController
