+++ app/frontend/entrypoints/france_map.js
@@ -1,0 +1,3 @@
+import ObjetsMapController from "../stimulus_controllers_standalone/france_map_controller"
+
+window.Stimulus.register("france-map", ObjetsMapController)
