--- app/frontend/entrypoints/objets_map.js
@@ -1,3 +1,0 @@
-import ObjetsMapController from "../stimulus_controllers_standalone/objets_map_controller"
-
-window.Stimulus.register("objets-map", ObjetsMapController)
