--- app/frontend/lib/openmaptiles_layers.js
@@ -1,2080 +1,0 @@
-export default [
-  {
-    "id": "background",
-    "type": "background",
-    "paint": {
-      "background-color": "#f8f4f0"
-    }
-  },
-  // {
-  //   "id": "water",
-  //   "type": "fill",
-  //   "metadata": {
-  //     "mapbox:group": "1444849382550.77"
-  //   },
-  //   "source": "openmaptiles",
-  //   "source-layer": "water",
-  //   "filter": [
-  //     "all",
-  //     [
-  //       "!=",
-  //       "intermittent",
-  //       1
-  //     ],
-  //     [
-  //       "!=",
-  //       "brunnel",
-  //       "tunnel"
-  //     ]
-  //   ],
-  //   "layout": {
-  //     "visibility": "visible"
-  //   },
-  //   "paint": {
-  //     "fill-color": "hsl(210, 67%, 85%)"
-  //   }
-  // },
-  {
-    "id": "highway-path",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "==",
-        "class",
-        "path"
-      ]
-    ],
-    "paint": {
-      "line-color": "#cba",
-      "line-dasharray": [
-        1.5,
-        0.75
-      ],
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            15,
-            1.2
-          ],
-          [
-            20,
-            4
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-motorway-link",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "minzoom": 12,
-    "filter": [
-      "all",
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "==",
-        "class",
-        "motorway"
-      ],
-      [
-        "==",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fc8",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12.5,
-            0
-          ],
-          [
-            13,
-            1.5
-          ],
-          [
-            14,
-            2.5
-          ],
-          [
-            20,
-            11.5
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-link",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "minzoom": 13,
-    "filter": [
-      "all",
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "in",
-        "class",
-        "trunk",
-        "primary",
-        "secondary",
-        "tertiary"
-      ],
-      [
-        "==",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round",
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12.5,
-            0
-          ],
-          [
-            13,
-            1.5
-          ],
-          [
-            14,
-            2.5
-          ],
-          [
-            20,
-            11.5
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-minor",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!=",
-        "brunnel",
-        "tunnel"
-      ],
-      [
-        "in",
-        "class",
-        "minor",
-        "service",
-        "track"
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fff",
-      "line-opacity": 1,
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            13.5,
-            0
-          ],
-          [
-            14,
-            2.5
-          ],
-          [
-            20,
-            11.5
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-secondary-tertiary",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "in",
-        "class",
-        "secondary",
-        "tertiary"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round",
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            6.5,
-            0
-          ],
-          [
-            8,
-            0.5
-          ],
-          [
-            20,
-            13
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-primary",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "in",
-        "class",
-        "primary"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round",
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            8.5,
-            0
-          ],
-          [
-            9,
-            0.5
-          ],
-          [
-            20,
-            18
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-trunk",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "in",
-        "class",
-        "trunk"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round",
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            6.5,
-            0
-          ],
-          [
-            7,
-            0.5
-          ],
-          [
-            20,
-            18
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "highway-motorway",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "minzoom": 5,
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "==",
-        "class",
-        "motorway"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round",
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "#fc8",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            6.5,
-            0
-          ],
-          [
-            7,
-            0.5
-          ],
-          [
-            20,
-            18
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "railway-transit",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "class",
-        "transit"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "tunnel"
-      ]
-    ],
-    "layout": {
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "hsla(0, 0%, 73%, 0.77)",
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14,
-            0.4
-          ],
-          [
-            20,
-            1
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "railway-transit-hatching",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "class",
-        "transit"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "tunnel"
-      ]
-    ],
-    "layout": {
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "hsla(0, 0%, 73%, 0.68)",
-      "line-dasharray": [
-        0.2,
-        8
-      ],
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14.5,
-            0
-          ],
-          [
-            15,
-            2
-          ],
-          [
-            20,
-            6
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "railway-service",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "class",
-        "rail"
-      ],
-      [
-        "has",
-        "service"
-      ]
-    ],
-    "paint": {
-      "line-color": "hsla(0, 0%, 73%, 0.77)",
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14,
-            0.4
-          ],
-          [
-            20,
-            1
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "railway-service-hatching",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "class",
-        "rail"
-      ],
-      [
-        "has",
-        "service"
-      ]
-    ],
-    "layout": {
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "hsla(0, 0%, 73%, 0.68)",
-      "line-dasharray": [
-        0.2,
-        8
-      ],
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14.5,
-            0
-          ],
-          [
-            15,
-            2
-          ],
-          [
-            20,
-            6
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "railway",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!has",
-        "service"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "==",
-        "class",
-        "rail"
-      ]
-    ],
-    "paint": {
-      "line-color": "#bbb",
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14,
-            0.4
-          ],
-          [
-            15,
-            0.75
-          ],
-          [
-            20,
-            2
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "railway-hatching",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "!has",
-        "service"
-      ],
-      [
-        "!in",
-        "brunnel",
-        "bridge",
-        "tunnel"
-      ],
-      [
-        "==",
-        "class",
-        "rail"
-      ]
-    ],
-    "paint": {
-      "line-color": "#bbb",
-      "line-dasharray": [
-        0.2,
-        8
-      ],
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14.5,
-            0
-          ],
-          [
-            15,
-            3
-          ],
-          [
-            20,
-            8
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-motorway-link-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "motorway"
-      ],
-      [
-        "==",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#e9ac77",
-      "line-opacity": 1,
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12,
-            1
-          ],
-          [
-            13,
-            3
-          ],
-          [
-            14,
-            4
-          ],
-          [
-            20,
-            19
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-link-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "trunk",
-        "primary",
-        "secondary",
-        "tertiary"
-      ],
-      [
-        "==",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#e9ac77",
-      "line-opacity": 1,
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12,
-            1
-          ],
-          [
-            13,
-            3
-          ],
-          [
-            14,
-            4
-          ],
-          [
-            20,
-            19
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-secondary-tertiary-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "secondary",
-        "tertiary"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#e9ac77",
-      "line-opacity": 1,
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            5,
-            0.4
-          ],
-          [
-            7,
-            0.6
-          ],
-          [
-            8,
-            1.5
-          ],
-          [
-            20,
-            21
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-trunk-primary-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "primary",
-        "trunk"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "hsl(28, 76%, 67%)",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            5,
-            0.4
-          ],
-          [
-            6,
-            0.6
-          ],
-          [
-            7,
-            1.5
-          ],
-          [
-            20,
-            26
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-motorway-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "motorway"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#e9ac77",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            5,
-            0.4
-          ],
-          [
-            6,
-            0.6
-          ],
-          [
-            7,
-            1.5
-          ],
-          [
-            20,
-            26
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-minor-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "minor",
-        "service",
-        "track"
-      ]
-    ],
-    "layout": {
-      "line-cap": "butt",
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#cfcdca",
-      "line-opacity": {
-        "stops": [
-          [
-            12,
-            0
-          ],
-          [
-            12.5,
-            1
-          ]
-        ]
-      },
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12,
-            0.5
-          ],
-          [
-            13,
-            1
-          ],
-          [
-            14,
-            6
-          ],
-          [
-            20,
-            24
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-path-casing",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "path"
-      ]
-    ],
-    "paint": {
-      "line-color": "#f8f4f0",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            15,
-            1.2
-          ],
-          [
-            20,
-            18
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-path",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "path"
-      ]
-    ],
-    "paint": {
-      "line-color": "#cba",
-      "line-dasharray": [
-        1.5,
-        0.75
-      ],
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            15,
-            1.2
-          ],
-          [
-            20,
-            4
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-motorway-link",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "motorway"
-      ],
-      [
-        "==",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fc8",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12.5,
-            0
-          ],
-          [
-            13,
-            1.5
-          ],
-          [
-            14,
-            2.5
-          ],
-          [
-            20,
-            11.5
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-link",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "trunk",
-        "primary",
-        "secondary",
-        "tertiary"
-      ],
-      [
-        "==",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            12.5,
-            0
-          ],
-          [
-            13,
-            1.5
-          ],
-          [
-            14,
-            2.5
-          ],
-          [
-            20,
-            11.5
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-minor",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849345966.4436"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "minor",
-        "service",
-        "track"
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fff",
-      "line-opacity": 1,
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            13.5,
-            0
-          ],
-          [
-            14,
-            2.5
-          ],
-          [
-            20,
-            11.5
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-secondary-tertiary",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "secondary",
-        "tertiary"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            6.5,
-            0
-          ],
-          [
-            8,
-            0.5
-          ],
-          [
-            20,
-            13
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-trunk-primary",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "in",
-        "class",
-        "primary",
-        "trunk"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fea",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            6.5,
-            0
-          ],
-          [
-            7,
-            0.5
-          ],
-          [
-            20,
-            18
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-motorway",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "motorway"
-      ],
-      [
-        "!=",
-        "ramp",
-        1
-      ]
-    ],
-    "layout": {
-      "line-join": "round"
-    },
-    "paint": {
-      "line-color": "#fc8",
-      "line-width": {
-        "base": 1.2,
-        "stops": [
-          [
-            6.5,
-            0
-          ],
-          [
-            7,
-            0.5
-          ],
-          [
-            20,
-            18
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-railway",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "rail"
-      ]
-    ],
-    "paint": {
-      "line-color": "#bbb",
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14,
-            0.4
-          ],
-          [
-            15,
-            0.75
-          ],
-          [
-            20,
-            2
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "bridge-railway-hatching",
-    "type": "line",
-    "metadata": {
-      "mapbox:group": "1444849334699.1902"
-    },
-    "source": "openmaptiles",
-    "source-layer": "transportation",
-    "filter": [
-      "all",
-      [
-        "==",
-        "brunnel",
-        "bridge"
-      ],
-      [
-        "==",
-        "class",
-        "rail"
-      ]
-    ],
-    "paint": {
-      "line-color": "#bbb",
-      "line-dasharray": [
-        0.2,
-        8
-      ],
-      "line-width": {
-        "base": 1.4,
-        "stops": [
-          [
-            14.5,
-            0
-          ],
-          [
-            15,
-            3
-          ],
-          [
-            20,
-            8
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "boundary-water",
-    "type": "line",
-    "source": "openmaptiles",
-    "source-layer": "boundary",
-    "minzoom": 4,
-    "filter": [
-      "all",
-      [
-        "in",
-        "admin_level",
-        2,
-        4
-      ],
-      [
-        "==",
-        "maritime",
-        1
-      ]
-    ],
-    "layout": {
-      "line-cap": "round",
-      "line-join": "round",
-      "visibility": "visible"
-    },
-    "paint": {
-      "line-color": "rgba(154, 189, 214, 1)",
-      "line-opacity": {
-        "stops": [
-          [
-            6,
-            0.6
-          ],
-          [
-            10,
-            1
-          ]
-        ]
-      },
-      "line-width": {
-        "base": 1,
-        "stops": [
-          [
-            0,
-            0.6
-          ],
-          [
-            4,
-            1.4
-          ],
-          [
-            5,
-            2
-          ],
-          [
-            12,
-            8
-          ]
-        ]
-      }
-    }
-  },
-  {
-    "id": "waterway-name",
-    "type": "symbol",
-    "source": "openmaptiles",
-    "source-layer": "waterway",
-    "minzoom": 13,
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "LineString"
-      ],
-      [
-        "has",
-        "name"
-      ]
-    ],
-    "layout": {
-      "symbol-placement": "line",
-      "symbol-spacing": 350,
-      "text-field": "{name:latin} {name:nonlatin}",
-      "text-font": [
-        "Noto Sans Italic"
-      ],
-      "text-letter-spacing": 0.2,
-      "text-max-width": 5,
-      "text-rotation-alignment": "map",
-      "text-size": 14
-    },
-    "paint": {
-      "text-color": "#74aee9",
-      "text-halo-color": "rgba(255,255,255,0.7)",
-      "text-halo-width": 1.5
-    }
-  },
-  {
-    "id": "water-name-lakeline",
-    "type": "symbol",
-    "source": "openmaptiles",
-    "source-layer": "water_name",
-    "filter": [
-      "==",
-      "$type",
-      "LineString"
-    ],
-    "layout": {
-      "symbol-placement": "line",
-      "symbol-spacing": 350,
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Italic"
-      ],
-      "text-letter-spacing": 0.2,
-      "text-max-width": 5,
-      "text-rotation-alignment": "map",
-      "text-size": 14
-    },
-    "paint": {
-      "text-color": "#74aee9",
-      "text-halo-color": "rgba(255,255,255,0.7)",
-      "text-halo-width": 1.5
-    }
-  },
-  {
-    "id": "water-name-ocean",
-    "type": "symbol",
-    "source": "openmaptiles",
-    "source-layer": "water_name",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "Point"
-      ],
-      [
-        "==",
-        "class",
-        "ocean"
-      ]
-    ],
-    "layout": {
-      "symbol-placement": "point",
-      "symbol-spacing": 350,
-      "text-field": "{name:latin}",
-      "text-font": [
-        "Noto Sans Italic"
-      ],
-      "text-letter-spacing": 0.2,
-      "text-max-width": 5,
-      "text-rotation-alignment": "map",
-      "text-size": 14
-    },
-    "paint": {
-      "text-color": "#74aee9",
-      "text-halo-color": "rgba(255,255,255,0.7)",
-      "text-halo-width": 1.5
-    }
-  },
-  {
-    "id": "water-name-other",
-    "type": "symbol",
-    "source": "openmaptiles",
-    "source-layer": "water_name",
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "Point"
-      ],
-      [
-        "!in",
-        "class",
-        "ocean"
-      ]
-    ],
-    "layout": {
-      "symbol-placement": "point",
-      "symbol-spacing": 350,
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Italic"
-      ],
-      "text-letter-spacing": 0.2,
-      "text-max-width": 5,
-      "text-rotation-alignment": "map",
-      "text-size": {
-        "stops": [
-          [
-            0,
-            10
-          ],
-          [
-            6,
-            14
-          ]
-        ]
-      },
-      "visibility": "visible"
-    },
-    "paint": {
-      "text-color": "#74aee9",
-      "text-halo-color": "rgba(255,255,255,0.7)",
-      "text-halo-width": 1.5
-    }
-  },
-  {
-    "id": "poi-railway",
-    "type": "symbol",
-    "source": "openmaptiles",
-    "source-layer": "poi",
-    "ST_GeomFromText('POINT(-71.064544 42.28787)');": 13,
-    "filter": [
-      "all",
-      [
-        "==",
-        "$type",
-        "Point"
-      ],
-      [
-        "has",
-        "name"
-      ],
-      [
-        "==",
-        "class",
-        "railway"
-      ],
-      [
-        "==",
-        "subclass",
-        "station"
-      ]
-    ],
-    "layout": {
-      "icon-allow-overlap": false,
-      "icon-ignore-placement": false,
-      "icon-image": "{class}_11",
-      "icon-optional": false,
-      "text-allow-overlap": false,
-      "text-anchor": "top",
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Regular"
-      ],
-      "text-ignore-placement": false,
-      "text-max-width": 9,
-      "text-offset": [
-        0,
-        0.6
-      ],
-      "text-optional": true,
-      "text-padding": 2,
-      "text-size": 12
-    },
-    "paint": {
-      "text-color": "#666",
-      "text-halo-blur": 0.5,
-      "text-halo-color": "#ffffff",
-      "text-halo-width": 1
-    }
-  },
-  {
-    "id": "place-village",
-    "type": "symbol",
-    "metadata": {
-      "mapbox:group": "1444849242106.713"
-    },
-    "source": "openmaptiles",
-    "source-layer": "place",
-    "filter": [
-      "==",
-      "class",
-      "village"
-    ],
-    "layout": {
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Regular"
-      ],
-      "text-max-width": 8,
-      "text-size": {
-        "base": 1.2,
-        "stops": [
-          [
-            10,
-            12
-          ],
-          [
-            15,
-            22
-          ]
-        ]
-      },
-      "visibility": "visible"
-    },
-    "paint": {
-      "text-color": "#333",
-      "text-halo-color": "rgba(255,255,255,0.8)",
-      "text-halo-width": 1.2
-    }
-  },
-  {
-    "id": "place-town",
-    "type": "symbol",
-    "metadata": {
-      "mapbox:group": "1444849242106.713"
-    },
-    "source": "openmaptiles",
-    "source-layer": "place",
-    "filter": [
-      "==",
-      "class",
-      "town"
-    ],
-    "layout": {
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Regular"
-      ],
-      "text-max-width": 8,
-      "text-size": {
-        "base": 1.2,
-        "stops": [
-          [
-            10,
-            14
-          ],
-          [
-            15,
-            24
-          ]
-        ]
-      },
-      "visibility": "visible"
-    },
-    "paint": {
-      "text-color": "#333",
-      "text-halo-color": "rgba(255,255,255,0.8)",
-      "text-halo-width": 1.2
-    }
-  },
-  {
-    "id": "place-city",
-    "type": "symbol",
-    "metadata": {
-      "mapbox:group": "1444849242106.713"
-    },
-    "source": "openmaptiles",
-    "source-layer": "place",
-    "filter": [
-      "all",
-      [
-        "!=",
-        "capital",
-        2
-      ],
-      [
-        "==",
-        "class",
-        "city"
-      ]
-    ],
-    "layout": {
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Regular"
-      ],
-      "text-max-width": 8,
-      "text-size": {
-        "base": 1.2,
-        "stops": [
-          [
-            7,
-            14
-          ],
-          [
-            11,
-            24
-          ]
-        ]
-      },
-      "visibility": "visible"
-    },
-    "paint": {
-      "text-color": "#333",
-      "text-halo-color": "rgba(255,255,255,0.8)",
-      "text-halo-width": 1.2
-    }
-  },
-  {
-    "id": "place-city-capital",
-    "type": "symbol",
-    "metadata": {
-      "mapbox:group": "1444849242106.713"
-    },
-    "source": "openmaptiles",
-    "source-layer": "place",
-    "filter": [
-      "all",
-      [
-        "==",
-        "capital",
-        2
-      ],
-      [
-        "==",
-        "class",
-        "city"
-      ]
-    ],
-    "layout": {
-      "icon-image": "star_11",
-      "icon-size": 0.8,
-      "text-anchor": "left",
-      "text-field": "{name:latin}\n{name:nonlatin}",
-      "text-font": [
-        "Noto Sans Regular"
-      ],
-      "text-max-width": 8,
-      "text-offset": [
-        0.4,
-        0
-      ],
-      "text-size": {
-        "base": 1.2,
-        "stops": [
-          [
-            7,
-            14
-          ],
-          [
-            11,
-            24
-          ]
-        ]
-      },
-      "visibility": "visible"
-    },
-    "paint": {
-      "text-color": "#333",
-      "text-halo-color": "rgba(255,255,255,0.8)",
-      "text-halo-width": 1.2
-    }
-  }
-]
