+++ app/views/departements/index.html.erb
@@ -1,6 +1,6 @@
 <% content_for(:head_title) { "Liste des communes" } %>
 
-<%= vite_javascript_tag "objets_map" %>
+<%= vite_javascript_tag "france_map" %>
 
 <main class="fr-container fr-pt-2w fr-pb-4w">
   <%= render(
@@ -17,13 +17,29 @@
   <div class="fr-grid-row fr-grid-row--gutters">
     <div class="fr-col-md-6 co-flex--grow">
       <div
-        data-controller="objets-map"
-        data-objets-map-target="container"
+        data-controller="france-map"
+        data-france-map-target="container"
+        data-departements-json="<%= @map_departements_json %>"
+        data-bins-json="<%= @map_bins.to_json %>"
         style="height: 600px; max-height: 50vh; width: 100%;">
       </div>
-      <p class="co-text--italic co-text--center">
+      <div class="co-text--italic co-text--center fr-my-1w">
         Nombre dʼobjets protégés par département et par commune
-      </p>
+      </div>
+      <div class="co-flex co-flex--justify-content-center">
+        <ul class="co-list-unstyled">
+          <% @map_bins.each do |bin| %>
+            <li>
+              <div
+                class="conservateur-map-legend-icon"
+                style="background-color: <%= bin[:color] %>"
+              >
+              </div>
+              <%= bin[:legend] %>
+            </li>
+          <% end %>
+        </ul>
+      </div>
     </div>
 
     <div class="fr-col-md-6">
